Fix MSVC C4244 in the fill journal guard
journalActive is a byte and its saved copy now is one too. The
jelly run's green was read from the wrong workflow run (a race with
the previous push); the completion release's own run caught it.
Lesson: verify CI runs by commit, not by recency.

+++ src/particle.c
@@ -823,7 +823,7 @@ int32_t m2World_FillPolygonWithParticles(m2WorldId worldId, const m2Polygon* pol
     // The whole fill is ONE journal op (inner emits suppressed, the
     // chain-create precedent): replay must rebuild the springs and
     // triads too, and those are captured here, not in the emits.
-    int32_t journalWas = world->journalActive;
+    uint8_t journalWas = world->journalActive;
     world->journalActive = 0;
 
     int32_t* emitted = (int32_t*)world->particleProxies; // borrowed scratch
